requirements.txt
alembic==1.13.1
blinker==1.6.3
certifi==2023.11.17
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
psycopg2==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.21
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==3.0.1
WTForms==3.0.1
zipp==3.17.0
